Clean up warnings.
Use (void)var to mark var as used, instead of var = var. The
intent of the self assignment was to avoid a warning, but it ends
up invoking a different warning.

Turn off security warnings for a potentially mutable string being
given as the format string to sprintf(). Though only on macosx
unix targets. All of the strings provided are in fact not mutable,
but this fact is hidden from the compiler due to the strings being
extern in a different object file.

(cherry picked from commit d685e65d71339cbdc0926b4cce072b6c15bb9b74)